+++ src/components/karaoke/PlaylistInfoBar.jsx
