+++ src/components/karaoke/player/Player.jsx
@@ -8,9 +8,9 @@ import { sendPlayerCommand } from 'actions/playlist'
 import { withNavigate } from 'components/adapted/ReactRouterDom'
 import { CSSTransitionLazy } from 'components/adapted/ReactTransitionGroup'
 import UserWidget from 'components/generics/UserWidget'
+import ManageButton from 'components/karaoke/player/ManageButton'
+import PlayerNotification from 'components/karaoke/player/Notification'
 import { IsPlaylistManagerOrOwner } from 'components/permissions/Playlist'
-import ManageButton from 'components/playlistApp/player/ManageButton'
-import PlayerNotification from 'components/playlistApp/player/Notification'
 import Song from 'components/song/Song'
 import { alterationResponsePropType, Status } from 'reducers/alterationsResponse'
 import { playlistDigestPropType } from 'reducers/playlist'
