+++ src/style/components/karaoke/kara_status_notification.scss
