+++ src/style/components/karaoke/player.scss
