+++ src/style/components/karaoke/playlist_info_bar.scss
